+++ Portal/src/Datahub.Infrastructure/Services/ResourceMessagingService.cs
@@ -4,6 +4,7 @@ using Datahub.Core.Model.Context;
 using Datahub.Core.Model.Datahub;
 using Datahub.Core.Utils;
 using Datahub.Infrastructure.Extensions;
+using Datahub.Shared;
 using Datahub.Shared.Configuration;
 using Datahub.Shared.Entities;
 using Datahub.Shared.Exceptions;
@@ -58,11 +59,13 @@ public class ResourceMessagingService(
 
         var workspace = project.ToResourceWorkspace(users);
         var templates = project.Resources
-            .Where(r => r.ResourceType != TerraformTemplate.VariableUpdate)
+            .Where(r => r.ResourceType != TerraformTemplate.VariableUpdate && !TerraformStatus.DeletedOrInProcessOf(r.Status))
             .Select(r => r.ToTerraformTemplate())
             .ToList();
 
-        workspace.Version = await workspaceVersionService.GetLatestVersion();
+
+        workspace.Version = workspace.Version == "latest" ? await workspaceVersionService.GetLatestVersion() : workspace.Version;
+
 
         return new WorkspaceDefinition
         {
@@ -76,5 +79,30 @@ public class ResourceMessagingService(
             },
             RequestingUserEmail = requestingUserEmail
         };
+    }
+
+    private string GetResourceNameSuffix(string templatetype, Datahub_Project project)
+    {
+        var resourceNumber = 0;
+
+        //get total resources of template type
+        switch (templatetype)
+        {
+            case TerraformTemplate.AzureAppService:
+                resourceNumber = project.Resources.Count(r => r.ResourceType.Equals(TerraformTemplate.GetTerraformServiceType(TerraformTemplate.AzureAppService)));
+                break;
+            case TerraformTemplate.AzurePostgres:
+                resourceNumber = project.Resources.Count(r => r.ResourceType.Equals(TerraformTemplate.GetTerraformServiceType(TerraformTemplate.AzurePostgres)));
+                break;
+            default:
+                throw new ArgumentException("Invalid template type", nameof(templatetype));
+        }
+
+        //get next iteration for suffix
+        resourceNumber++;
+
+        // format resourceNumber to three digits
+        return resourceNumber.ToString("D3");
+
     }
 }
